// src/typings/user.d.ts
export type MemberVerifyMemberByAccounReq = {
  type: number // 类型 1 手机 2 邮箱 3 uid
  account: string // 邮箱
  password?: string // 密码
  mobileCountryCode: string // 区号
}

export type MemberVerifyMemberByAccounResp = {
  isOpenEmailVerify: number // 是否启用邮箱 1 启用 2 未启用
  isOpenGoogleVerify: number // 是否启用谷歌 1 启用 2 未启用
  isOpenPhoneVerify: number // 是否启用手机 1 启用 2 未启用
  mobileCountryCd?: string // 手机区号
  mobileNumber?: string // 手机号
  kycType: number // kyc 等级 1 为未认证 2 为个人标准认证 3 为个人高级认证 4 为企业认证
  email?: string // 邮箱
}